refactor: Update import path for Navigation component
The code changes in the `App.jsx` file update the import path for the `Navigation` component from `'./pages/Navigation'` to `'./components/Navigation'`. This change ensures that the correct component is imported and used in the application.

Based on the recent repository commits, the commit message follows the convention of starting with a type (`refactor` for code refactoring) and providing a concise description of the changes made.
